add un po' di funzioni mancanti sui gruppi

+++ BoardGameApp/src/main/java/org/openjfx/DBManager/Neo4jDBManager/GamesReviewsRatesDBManager.java
@@ -128,14 +128,14 @@ public class GamesReviewsRatesDBManager {
         return numberRates;
     }
 
-    public static float avgRates(String name)
+    public static Double avgRates(String name)
     {
         try(Session session=driver.session())
         {
-            return session.writeTransaction(new TransactionWork<Integer>()
+            return session.writeTransaction(new TransactionWork<Double>()
             {
                 @Override
-                public Integer execute(Transaction tx)
+                public Double execute(Transaction tx)
                 {
                     return transactionAvgRates(tx, name);
                 }
@@ -145,16 +145,16 @@ public class GamesReviewsRatesDBManager {
 
     //Funzione che fa la media delle valutazioni ad un determinato gioco
 
-    public static int transactionAvgRates(Transaction tx, String name) {
+    public static Double transactionAvgRates(Transaction tx, String name) {
 
-        int avgRates = 0;
+        Double avgRates = 0.0;
         HashMap<String, Object> parameters = new HashMap<>();
         parameters.put("name", name);
         Result result = tx.run("MATCH (ul:User)-[r:RATED]->(g:Game{name:$name}) return avg(r.vote) AS avgRates", parameters);
 
         if (result.hasNext()) {
             Record record = result.next();
-            avgRates = record.get("avgRates").asInt();
+            avgRates = record.get("avgRates").asDouble();
 
         }
         return avgRates;
